+++ NEWS.md
@@ -1,6 +1,7 @@
 ### HEAD (unreleased)
 
 * Massive refactoring
+* Chat times respect time zones instead of assuming inputs are UTC
 
 ### 3.3.0 / 2011-10-16
 * Pidgin2Adium depended on itself for some reason. Now it doesn't.